@CeruleanSky 2024-04-10
I know this is closed, but for reference purposes I ran your micro benchmark on an intel i7-7700HQ several times and here is a record of the typical values

Bench(0xD6E93);
```
fast rsqrte took 8 ms
rsqrte took 24 ms
rsqrt took 8 ms
```
Bench(0.42f);
```
fast rsqrte took 5 ms
rsqrte took 23 ms
rsqrt took 10 ms
```