@gdkchan 2024-04-10
I made a mistake on the benchmark code where I used `MathF.Sqrt` instead of `Sse.SqrtScalar` which made the `rsqrte` results slower than it should be. I have corrected it now and updated the resulst. It does not change much here, it is still what one would expect with `fast rsqrte` being the fastest, `rsqrt` the slowest and `rsqrte` being in the middle.