@CeruleanSky 2024-04-10
I updated the times with the correction. `rsqrte` is always seems faster for Intel for the subnormal case than `fast rsqrte` for the 4M rounds.  Seems in the middle by a small margin when upped to 400M rounds.